% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancesCalc.R
\name{addDistanceSpeed}
\alias{addDistanceSpeed}
\title{calculate distance at space and time columns}
\usage{
addDistanceSpeed(input.loc.df, merics.vec = c("distance", "dT", "spd"))
}
\arguments{
\item{input.loc.df}{localization data frame}

\item{merics.vec}{which metrics to add (distance, dT, spd)}
}
\value{
input.loc.df
}
\description{
TODO...
}
\examples{
TODO
}
