% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movementsData.R
\name{addLocAttribute}
\alias{addLocAttribute}
\title{add indexes uses for filtering by loclization quality}
\usage{
addLocAttribute(.df, locAttributs = c("distanceSpeed", "angle", "locQuality"))
}
\arguments{
\item{.df}{raw atlas data.frame}

\item{locAttributs}{vectore of attibutes per localizations record
<ul>
<li>"distanceSpeed" or "distance" or "speed" -  add 3 attibutes: dT, distances and speed</li>
<li>"angle" - calculte angle between 2 successive path (-180..180 degrees)</li>
<li>"locQuality" - calculate 2 localization quality attibutes
               <li>stdErrXY - based on COVX, VARX, VARY</li>
               <li>traceNorm - based o, VARX and VARY</li></li>
</ul>
default : all attributes}
}
\value{
data.frame with added columns
}
\description{
TODO...
}
\examples{
addLocAttribute(df)
addLocAttribute(df, c("distanceSpeed", "locQuality"))

}
