% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrameTimeFractions.r
\name{addTimeDateFields}
\alias{addTimeDateFields}
\title{add time categories for further aggregate atlas data (locallizations/detections) data}
\usage{
addTimeDateFields(df, fieldsTypes, refTime = 0)
}
\arguments{
\item{df}{data frame whith a column names "TIME" , its values are epoch time in milisec resolution in UTCtime zone}

\item{fieldsTypes}{list of time fields types text:
- "date"
- "hour", "min"
- "day"
@param refTime - reference time (default 0)}
}
\value{
data.frame with added columns in format <timeUnit><number>
}
\description{
TODO...
}
\examples{
TODO

}
