% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdpFixedPoint.R
\name{AdpFixedPoint}
\alias{AdpFixedPoint}
\title{AdpFixedPoint is a function to perform track segmentation utilizing a
first-passage algorithm to determine fixed points where the agent/animal
has spent a minimum number of observations (obs_min) within a a certain
range (adp_rng) of a fixed point that is continuously reavaluated.}
\usage{
AdpFixedPoint(
  time.vec,
  x,
  y,
  adp_rng,
  smp_rte,
  obs_min,
  p_lim,
  time_gap = 5 * 60 * 1000
)
}
\arguments{
\item{time.vec}{absolute time e.g. in ms (ATLAS timestamp)}

\item{x}{projected longitude in meters}

\item{y}{projected latitude in meters}

\item{adp_rng}{adaptive range defining fixed points}

\item{smp_rte}{sampling rate e.g. in ms (ATLAS)}

\item{obs_min}{minimum nuber of observations defining a fixed position}

\item{p_lim}{point limit for leaving current fixed point}

\item{time_gap}{minimal time gap of missing values refere as exit of adp area (default 5 minutes)}
}
\value{
The function returns an array containing information about each fixed
        point including (in order) start time, end time, duration, number of locations,
        position quality, median x, median-y, lower x quantile, upper x quantile, lower y quantile, upper y quantile


Original Code by Ingo Schiffner 2017 (Matlab version)
Converted to R by Emanuel Lourie 2018
}
\description{
AdpFixedPoint is a function to perform track segmentation utilizing a
first-passage algorithm to determine fixed points where the agent/animal
has spent a minimum number of observations (obs_min) within a a certain
range (adp_rng) of a fixed point that is continuously reavaluated.
}
