% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagsProperties.R
\name{calcTagRate}
\alias{calcTagRate}
\title{define tags broadcast  rate}
\usage{
calcTagRate(tags.lst, timeunit = "ms")
}
\arguments{
\item{tags.lst}{input data frame contains TAG, TIME}

\item{timeUnit}{timestampe units. default : ms (means milisec)}
}
\value{
data frame with tag name and freq
}
\description{
define tags broadcast  rate
}
\examples{
TODO


}
