% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackAngles.R
\name{cleanShrpFastPath}
\alias{cleanShrpFastPath}
\title{clean Shrp angle with quick movement - it not realistic}
\usage{
cleanShrpFastPath(df, anglTh = 10, distTh = 100, timeFrame)
}
\arguments{
\item{df}{one data frame}

\item{anglTh}{sngle threshold}

\item{distTh}{distance threshold}

\item{timeFrame}{timeframe}
}
\value{
cleaned data frame
}
\description{
TODO...
}
\examples{
TODO

}
