% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVis.R
\name{convert.sf.ITM2WGS84}
\alias{convert.sf.ITM2WGS84}
\title{convert data.frame with ITM coordinates to sf for WGS84
using crs=2039}
\usage{
convert.sf.ITM2WGS84(df, xyColNames = c("X", "Y"))
}
\arguments{
\item{df}{input data frame}

\item{xyColNames}{vector with the nameof the X, Y column (by ITM)}
}
\value{
spatial data frame for WGS84
}
\description{
convert data.frame with ITM coordinates to sf for WGS84
using crs=2039
}
\examples{
TODO

}
