% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVis.R
\name{convertSpatial.WGS84}
\alias{convertSpatial.WGS84}
\title{convert data.frame with WGS coordinates to spatiall data.frame for ITM}
\usage{
convertSpatial.WGS84(wgs84.df, xyColNames = c("LON", "LAT"))
}
\arguments{
\item{wgs84.df}{input data frame}

\item{xyColNames}{vector with the nameof the LON, LOT column (by WGS84)}
}
\value{
spatial data frame for ITM
}
\description{
convert data.frame with WGS coordinates to spatiall data.frame for ITM
}
\examples{
TODO

}
