% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movementsData.R
\name{defineSectionsByTime}
\alias{defineSectionsByTime}
\title{mark sections of sequence localizations}
\usage{
defineSectionsByTime(.df, dtThreshold = 600)
}
\arguments{
\item{dtThreshold}{timethreshold (in seconds)}

\item{df}{localization data frame}
}
\value{
data.frame with added columns sectionID
}
\description{
TODO...
}
\examples{
defineSectionsByTime(df, 600) #10 mintes gap seperate sections

}
