% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancesCalc.R
\name{distances.spd.Calc}
\alias{distances.spd.Calc}
\title{calculate distance between sequence locations}
\usage{
distances.spd.Calc(x, y, t)
}
\arguments{
\item{x}{x vector}

\item{y}{y vector}

\item{t}{time vector}
}
\value{
dataframe
}
\description{
TODO...
}
\examples{
TODO

}
