% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByVelocity.R
\name{filterByVelocity}
\alias{filterByVelocity}
\title{filter by velocity}
\usage{
filterByVelocity(
  locs,
  dataLegend = c("X", "Y", "TIME"),
  options = c(v_filter_max_v = 20, min_time_dif_for_stats = 10, v_filter_time_dif = 10)
)
}
\arguments{
\item{df}{one data frame}

\item{anglTh}{sngle threshold}

\item{distTh}{distance threshold}

\item{timeFrame}{timeframe}
}
\value{
cleaned data frame
}
\description{
TODO...
}
\examples{
TODO

}
