% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedSmooth.R
\name{MedConvSmoothing}
\alias{MedConvSmoothing}
\title{median convolution
calculates median x y localizations and time (t) over a predefined window (win) in ms
assumes x and y coordinates are given in a projected format in meters and time(t) given as ATLASTimeStamp(ms)
returns a data.frame containing x,y and time(t) and cnt is number of localizations used for median calculation}
\usage{
MedConvSmoothing(x, y, t, win, freq, skip = 1)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{t}{time}

\item{win}{window (ms)}

\item{freq}{(ms)}

\item{skip}{(ms) default =1}
}
\value{
data frame with median values
}
\description{
median convolution
calculates median x y localizations and time (t) over a predefined window (win) in ms
assumes x and y coordinates are given in a projected format in meters and time(t) given as ATLASTimeStamp(ms)
returns a data.frame containing x,y and time(t) and cnt is number of localizations used for median calculation
}
\examples{
TODO

}
