% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedSmooth.R
\name{MedSmoothing}
\alias{MedSmoothing}
\title{basic median filter by Ingo Schiffner 2017
calculates median x y localizations and time (t) over a predefined window (win) in ms
assumes x and y coordinates are given in a projected format in meters and time(t) given as ATLASTimeStamp(ms)
returns a data.frame containing x,y and time(t) and cnt is number of localizations used for median calculation}
\usage{
MedSmoothing(x, y, t, win)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{t}{time}

\item{win}{window}
}
\value{
data frame with median values
}
\description{
basic median filter by Ingo Schiffner 2017
calculates median x y localizations and time (t) over a predefined window (win) in ms
assumes x and y coordinates are given in a projected format in meters and time(t) given as ATLASTimeStamp(ms)
returns a data.frame containing x,y and time(t) and cnt is number of localizations used for median calculation
}
\examples{
TODO


}
