% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det-loc.R
\name{mergeLocDet}
\alias{mergeLocDet}
\title{merge BS from detections from localizations}
\usage{
mergeLocDet(loc.df, det.df)
}
\arguments{
\item{loc.df}{data frame of atlas localizations (should include at least TAG, TIME, X,Y)}

\item{det.df}{data frame of atlas detections (should include at least TAG, TIME, BS)}
}
\value{
merged table
}
\description{
TODO...
}
\examples{
loadFromSQLite(<sqliteFieName>)

}
