% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqliteInterface.R
\name{saveIntoSQLite}
\alias{saveIntoSQLite}
\title{write to ATLAS sqlite tables}
\usage{
saveIntoSQLite(dbn, tbls, append = FALSE)
}
\arguments{
\item{dbn}{sqlite file name}

\item{tbl}{table name}
}
\description{
TODO...
}
\examples{
saveIntoSQLite(<sqliteFieName>)

}
