% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticDesc.R
\name{showHistograms}
\alias{showHistograms}
\title{showHistograms}
\usage{
showHistograms(
  df,
  histList = c("spd", "traceNorm", "dT"),
  groupBy = c("TAG"),
  txt = "without filter"
)
}
\arguments{
\item{df}{atlas data.frame with parameters for histogram}

\item{histList}{list of field for histogram
default : spd, traceNorm, dT}

\item{groupBy}{list of categories field to group by}

\item{txt}{for annotation title [not works yet]}
}
\value{
list of figures
}
\description{
plot histograms for list of parameters - each figure has place for 4 plots
}
\examples{
showHistograms(df, histList=c("spd","traceNorm","dT"), groupBy=c("TAG"))

}
