% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdvFilt.R
\name{stdevFilt}
\alias{stdevFilt}
\title{calculate stdErrXY}
\usage{
stdevFilt(varX, varY, covXY)
}
\arguments{
\item{varX}{X varance  vector}

\item{varY}{Y vriance vector}

\item{covXY}{common variance vector}
}
\value{
vector of formula for accuracy
}
\description{
TODO...
}
\examples{
TODO
}
