% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterVis.R
\name{tag.filters.ll.map}
\alias{tag.filters.ll.map}
\title{create maps with filters layers to select the best filter for you}
\usage{
tag.filters.ll.map(t_dat.spdf, filters.df, KalmanFilter.spdf = NULL)
}
\arguments{
\item{t_dat.spdf}{spatial point data frame with your data (should contains covXYGroup)}

\item{filters.df}{data fame define some filters to investigate}

\item{KalmanFilter.spdf}{dpatial point for kalman smothing (default NULL)}
}
\value{
leaflet map
}
\description{
TODO...
}
\examples{
TODO
}
