% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackAngles.R
\name{TrackAngles}
\alias{TrackAngles}
\title{calculate anglebetween each 3 loclizations}
\usage{
TrackAngles(x, y)
}
\arguments{
\item{x}{vector x}

\item{y}{vector y}
}
\value{
vector of angles
}
\description{
TODO...
}
\examples{
TODO

}
