% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps2atlas.R
\name{convertSpatial.ITM2WGS84}
\alias{convertSpatial.ITM2WGS84}
\title{convert from ITM to GWS84 (data frame)}
\usage{
convertSpatial.ITM2WGS84(df, xyColNames = c("X", "Y"))
}
\arguments{
\item{df}{input data frame}

\item{xyColNames}{vector with the nameof the X, Y column (by ITM)}
}
\value{
spatial data frame for WGS84
}
\description{
convert data.frame with ITM coordinates to spatiall data.frame for WGS84
}
\examples{
TODO

}
