% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqliteInterface.R
\name{loadFromSQLite}
\alias{loadFromSQLite}
\title{interface to ATLAS sqlite tables}
\usage{
loadFromSQLite(dbn, tbl = "localizations")
}
\arguments{
\item{dbn}{sqlite file name}

\item{tbl}{table name}
}
\value{
table content as data.frame
}
\description{
TODO...
}
\examples{
loadFromSQLite(<sqliteFieName>)

}
