% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticDesc.R
\name{statistic.desc}
\alias{statistic.desc}
\title{statistics describe parameter disrtribution}
\usage{
statistic.desc(df, parameter = NULL, groupBy = NULL)
}
\arguments{
\item{df}{atlas data.frame with parameters for histogram}

\item{parameter}{parameer's field name
default : NULL}

\item{groupBy}{list of categories field to group by
default : NULL}
}
\value{
data frame of statistic information
}
\description{
statistics describe parameter disrtribution
}
\examples{
statistic.desc(df, parameter="spd", groupBy=c("TAG"))

}
